\name{testprint}
\title{Perform Factor Analysis}
\usage{
FACTOR(data)
}
\description{
This function performs factor analysis on a dataset.
}


#' @name testprint
#' @title Perform Factor Analysis
#' @description This function performs factor analysis on a dataset.
#' @param data A data frame containing the variables to be analyzed.
#' @return A list containing the results of the factor analysis.
#' @export

testprint <- function(x) {
  print(c("Hello, world!", x))
}


